requirements.txt
# ========================================
#   Weibo輿情分析多智能體系統依賴包
#   適用於Windows環境 + Conda部署
# ========================================

# ===== 核心Web框架 =====
flask==2.3.3
flask-socketio==5.3.6
streamlit==1.28.1
python-socketio==5.8.0
eventlet==0.33.3

# ===== HTTP請求和異步 =====
requests==2.31.0
httpx==0.28.1
aiofiles==23.2.1
aiohttp>=3.8.0

# ===== LLM接口 =====
openai>=1.3.0
# deepseek-ai>=0.1.0  # 使用OpenAI格式

# ===== 搜索API =====
tavily-python>=0.3.0

# ===== 數據處理 =====
pandas>=2.0.0
numpy>=1.24.0
regex>=2023.8.8
jieba==0.42.1

# ===== 數據庫 =====
pymysql==1.1.0
aiomysql==0.2.0
aiosqlite==0.21.0
redis>=4.6.0
SQLAlchemy==2.0.35
asyncpg==0.29.0

# ===== 爬蟲相關 =====
playwright==1.45.0
Pillow==9.5.0
opencv-python>=4.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
parsel==1.9.1
pyexecjs==1.5.1

# ===== 可視化 =====
plotly>=5.17.0
matplotlib==3.9.0
wordcloud==1.9.3

# ===== 機器學習（可選，用於情感分析，不安裝也沒事寫了容錯程序） =====
# torch>=2.0.0 # CPU版本 (已改用 GPU 版本，在 Dockerfile 中安裝)
# GPU版本: pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu126
transformers>=4.30.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# ===== 工具庫 =====
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0
tenacity==8.2.2
loguru>=0.7.0
pydantic==2.5.2
pydantic-settings==2.2.1

# ===== 開發工具（可選） =====
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# ===== Web服務器 =====
fastapi==0.110.2
uvicorn==0.29.0